refactor(agents): promote the skill-library mixin to the framework, fix the sidecar's code-index root
SkillLibraryToolsMixin imported nothing gaia-agent-specific — it is general
'let the model manage its own skill library' capability, and its four trust
refusals are framework decisions a second agent would have re-derived. It
now lives in gaia.agents.tools.skill_library_tools, is composable by name
(KNOWN_TOOLS['skills']), and the old import path re-exports it.

GaiaAgent's code index was rooted at os.getcwd() — the same wrong cwd its
allowed_paths field documents rejecting (the daemon launches the sidecar
with cwd = the package directory), so code search was sandboxed to the
agent's own source tree and indexed it by default. It now shares
allowed_paths' scope. The base class also moves first in the MRO, matching
every other hub agent so a future mixin method can't silently win over
ChatAgent's.

setup.py's mcp comment is aligned with the block main merged in #2941, so
the branches no longer conflict on that hunk.
